Sql intro2.0

+++ src/main/java/pro/sky/hogwarts/service/impl/FacultyServiceImpl.java
@@ -1,20 +1,24 @@
 package pro.sky.hogwarts.service.impl;
 
 import org.springframework.stereotype.Service;
+import pro.sky.hogwarts.Exception.FacultyNotFoundException;
 import pro.sky.hogwarts.model.Faculty;
 import pro.sky.hogwarts.model.Students;
 import pro.sky.hogwarts.repository.FacultyRepository;
+import pro.sky.hogwarts.repository.StudentRepository;
 import pro.sky.hogwarts.service.FacultyService;
 
-import java.util.Collection;
+import java.util.List;
 
 @Service
 public class FacultyServiceImpl implements FacultyService {
 
     FacultyRepository facultyRepository;
+    StudentRepository studentRepository;
 
-    public FacultyServiceImpl(FacultyRepository facultyRepository) {
+    public FacultyServiceImpl(FacultyRepository facultyRepository, StudentRepository studentRepository) {
         this.facultyRepository = facultyRepository;
+        this.studentRepository = studentRepository;
     }
 
     @Override
@@ -28,22 +32,33 @@ public class FacultyServiceImpl implements FacultyService {
     }
 
     @Override
-    public Collection<Faculty> getAllFaculties() {
+    public List<Faculty> getAllFaculties() {
         return facultyRepository.findAll();
     }
 
     @Override
     public Faculty getFacultyById(Long id) {
-        return facultyRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("Факультет не найден"));
+        return facultyRepository.findById(id).orElseThrow(() -> new FacultyNotFoundException("Faculty with id: " + id + "not found"));
     }
 
     @Override
-    public Collection<Faculty> findFaculty(String strSearch) {
+    public List<Faculty> findFaculty(String strSearch) {
         return facultyRepository.findByNameContainingIgnoreCaseOrColorContainingIgnoreCase(strSearch, strSearch);
     }
 
     @Override
     public void deleteFaculty(Long id) {
+        if (!facultyRepository.existsById(id)) {
+            throw new FacultyNotFoundException("Faculty with id: " + id + " not found");
+        }
         facultyRepository.deleteById(id);
+    }
+
+    @Override
+    public List<Students> getStudentsByFacultyId(Long id) {
+        if (!facultyRepository.existsById(id)) {
+            throw new FacultyNotFoundException("Faculty with id: " + id + "not found");
+        }
+        return studentRepository.findAllByFacultyId(id);
     }
 }
